# src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/ReservationDto.java
package ar.edu.utn.frc.tup.lc.iv.dtos;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;
@Data
@NoArgsConstructor
public class ReservationDto {
    private String id;
    private String status;
    private String flight;
    private List<PassengersDto> passengers;
}